Move getSimpleName() utility method to a utility class.
Eliminates a direct dependency on the diagnostic collectors from the quick fix classes.
